refactor(api): spilt code into different files
