+++ src/app/api/requests/account.ts
@@ -1,14 +1,14 @@
-import { User, UserFormValues } from '../../models/user';
-import { requests } from './base';
+import { User, UserFormValues } from "../../models/user";
+import { requests } from "./base";
 
 export const Account = {
-  current: () => requests.get<User>('/account'),
-  login: (user: UserFormValues) => requests.post<User>('/account/login', user),
+  current: () => requests.get<User>("/account"),
+  login: (user: UserFormValues) => requests.post<User>("/account/login", user),
   register: (user: UserFormValues) =>
-    requests.post<User>('/account/register', user),
+    requests.post<User>("/account/register", user),
   fbLogin: (accessToken: string) =>
     requests.post<User>(`/account/fbLogin?accessToken=${accessToken}`, {}),
-  refreshToken: () => requests.post<User>('/account/refreshToken', {}),
+  refreshToken: () => requests.post<User>("/account/refreshToken", {}),
   verifyEmail: (token: string, email: string) =>
     requests.post<void>(
       `/account/verifyEmail?token=${token}&email=${email}`,
